Update directory of news page to blogs
